requirement.txt
numpy==1.21.6
pandas==1.3.5
streamlit
pickle
scikit-learn
keras_model.h5
pip==23.1.2